requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-parler==2.3
django-parler-rest==2.2
django-phonenumber-field==7.2.0
django-redis==5.4.0
djangorestframework==3.14.0
executing==2.0.1
gunicorn==22.0.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==24.0
parso==0.8.3
pexpect==4.9.0
phonenumbers==8.13.25
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytz==2023.3.post1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
urllib3==2.1.0
wcwidth==0.2.13
